requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.5
Werkzeug==3.0.1
APScheduler==3.10.4
python-dotenv==1.0.0
email-validator==2.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

